# server/openapi/giantpay-v1.yaml
openapi: 3.1.0
info:
  title: GiantPay API
  version: 0.1.0
  description: |
    Contract for the GiantPay backend foundation. Amounts are integer minor units. Sandbox only.
    Cookie-authenticated mutations require X-CSRF-Token; bearer API-key requests do not.
    Credentialed browser requests are accepted only from the configured CORS origin.
    Rate-limited responses include Retry-After, RateLimit-Limit, RateLimit-Remaining and RateLimit-Reset.
servers:
  - url: http://127.0.0.1:4000/v1
...
paths:
  /notifications:
    get:
      operationId: listNotifications
      tags: [Notifications]
      summary: List the authenticated merchant user's in-app notifications
      parameters:
        - { in: query, name: cursor, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses: { '200': { description: User-isolated notification page }, '401': { description: Authentication required }, '403': { description: Permission denied } }
...
  /notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [Notifications]
      summary: Get effective notification preferences and channel availability
      responses: { '200': { description: Effective preferences } }
    patch:
      operationId: updateNotificationPreference
      tags: [Notifications]
      summary: Update one versioned notification preference
      parameters: [{ in: header, name: Idempotency-Key, required: true, schema: { type: string, minLength: 8, maxLength: 128 } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [category, channel, enabled, expectedVersion], properties: { category: { type: string }, channel: { enum: [IN_APP, EMAIL, SMS] }, enabled: { type: boolean }, expectedVersion: { type: integer, minimum: 0 } } } } } }
      responses: { '200': { description: Updated preference; external delivery remains unavailable }, '409': { description: 'Mandatory policy, stale version, or idempotency conflict' } }
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [Notifications]
      summary: Idempotently mark an owned notification read
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }, { in: header, name: Idempotency-Key, required: true, schema: { type: string } }]
      responses: { '200': { description: Original or replayed read result }, '404': { description: Notification not found } }
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [Notifications]
      summary: Idempotently mark the authenticated user's notifications read
      parameters: [{ in: header, name: Idempotency-Key, required: true, schema: { type: string } }]
      responses: { '200': { description: Read count } }
...
  /platform/notifications/{id}/read:
    post:
      operationId: markPlatformNotificationRead
      tags: [Platform notifications]
      summary: Mark an owned platform notification read
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }, { in: header, name: Idempotency-Key, required: true, schema: { type: string } }]
      responses: { '200': { description: Read result }, '404': { description: Notification not found } }
  /platform/notifications/read-all:
    post:
      operationId: markAllPlatformNotificationsRead
      tags: [Platform notifications]
      summary: Mark all notifications for the authenticated platform user read
      parameters: [{ in: header, name: Idempotency-Key, required: true, schema: { type: string } }]
      responses: { '200': { description: Read count } }
...
components:
  securitySchemes:
    cookieSession: { type: apiKey, in: cookie, name: giantpay_session }
    sandboxApiKey: { type: http, scheme: bearer, bearerFormat: gp_test_<identifier>_<secret> }
...