# server/src/notifications/routes.ts
import { createHash } from 'node:crypto';
import type { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
import { z } from 'zod';
import type { Config } from '../config.js';
import type { Db } from '../db.js';
import { transaction } from '../db.js';
import { rateLimit, type RateLimitStore } from '../rateLimit.js';
import { apiError, authenticate, newId } from '../security.js';
import { operationalControlGuard } from '../operations/controls.js';

const categories=['SECURITY','PAYMENT','REFUND','SETTLEMENT','RECONCILIATION','COMPLIANCE','ONBOARDING','SUPPORT','DISPUTE','TEAM_ACCESS','SYSTEM'] as const;
const channels=['IN_APP','EMAIL','SMS'] as const;
const mandatory=new Set(['SECURITY','COMPLIANCE','TEAM_ACCESS']);
const key=z.string().min(8).max(128),id=z.string().min(1).max(200);
const canonical=(x:any):string=>x===null||typeof x!=='object'?JSON.stringify(x):Array.isArray(x)?`[${x.map(canonical).join(',')}]`:`{${Object.keys(x).sort().map(k=>`${JSON.stringify(k)}:${canonical(x[k])}`).join(',')}}`;
const digest=(x:unknown)=>createHash('sha256').update(canonical(x)).digest('hex');
const cursor=(value?:string)=>{if(!value)return [null,null] as const;try{const [at,rowId]=JSON.parse(Buffer.from(value,'base64url').toString());if(!at||!rowId||Number.isNaN(Date.parse(at)))throw 0;return [at,rowId] as const;}catch{throw new NotificationError('NOTIFICATION_CURSOR_INVALID','The cursor is invalid.',400)}};
const next=(row:any)=>Buffer.from(JSON.stringify([row.created_at,row.id])).toString('base64url');

class NotificationError extends Error{constructor(public code:string,message:string,public status=409){super(message)}}
const fail=(r:FastifyRequest,p:FastifyReply,e:unknown)=>e instanceof NotificationError?p.code(e.status).send(apiError(r,e.code,e.message)):Promise.reject(e);
const view=(x:any)=>({id:x.id,category:x.category,templateKey:x.template_key,templateVersion:Number(x.template_version),title:x.title,body:x.body,channel:x.channel,deliveryState:x.delivery_state,relatedEntityType:x.related_entity_type,relatedEntityId:x.related_entity_id,createdAt:x.created_at,availableAt:x.available_at,expiresAt:x.expires_at,readAt:x.read_at??null});
const human=(db:Db,config:Config,authority:'MERCHANT'|'PLATFORM',permission:string)=>async(r:FastifyRequest,p:FastifyReply)=>{if(r.headers.authorization)return p.code(403).send(apiError(r,'FORBIDDEN','A browser session is required.'));await authenticate(db,config,r,p);if(p.sent)return;if(authority==='MERCHANT'&&r.actor?.merchantId===null)return p.code(403).send(apiError(r,'FORBIDDEN','Merchant authority is required.'));if(authority==='PLATFORM'&&(r.actor?.merchantId!==null||r.actor?.role!=='PLATFORM_ADMIN'))return p.code(403).send(apiError(r,'PLATFORM_PERMISSION_DENIED','Platform staff authority is required.'));if(!r.actor?.permissions.includes(permission))return p.code(403).send(apiError(r,authority==='PLATFORM'?'PLATFORM_PERMISSION_DENIED':'FORBIDDEN','You do not have permission to perform this action.'));};
async function idem(c:any,r:FastifyRequest,operation:string,resource:string,input:unknown,work:()=>Promise<{status:number,body:any}>){const k=key.parse(r.headers['idempotency-key']),authority=r.actor!.merchantId?'MERCHANT':'PLATFORM',hash=digest(input);await c.query('SELECT pg_advisory_xact_lock(hashtext($1))',[`notification:${authority}:${r.actor!.id}:${operation}:${resource}:${k}`]);const old=(await c.query('SELECT * FROM notification_idempotency WHERE authority_scope=$1 AND actor_id=$2 AND operation=$3 AND resource_id=$4 AND idempotency_key=$5',[authority,r.actor!.id,operation,resource,k])).rows[0];if(old){if(old.request_sha256!==hash)throw new NotificationError('IDEMPOTENCY_CONFLICT','The idempotency key was reused with different input.');return {status:old.response_status,body:old.response_body};}const out=await work();await c.query('INSERT INTO notification_idempotency(authority_scope,actor_id,operation,resource_id,idempotency_key,request_sha256,response_status,response_body) VALUES($1,$2,$3,$4,$5,$6,$7,$8)',[authority,r.actor!.id,operation,resource,k,hash,out.status,out.body]);return out;}
async function evidence(c:any,r:FastifyRequest,action:string,resource:string,metadata:object={}){await c.query(`INSERT INTO audit_events(id,actor_id,merchant_id,action,resource_type,resource_id,metadata) VALUES($1,$2,$3,$4,'notification',$5,$6)`,[newId('aud'),r.actor!.id,r.actor!.merchantId,action,resource,metadata]);await c.query(`INSERT INTO outbox_events(id,event_type,aggregate_type,aggregate_id,payload,deduplication_key) VALUES($1,$2,'notification',$3,$4,$5)`,[newId('obx'),action.toLowerCase().replaceAll('_','.'),resource,{notificationId:resource,merchantId:r.actor!.merchantId,...metadata},`${action}:${resource}:${newId('evt')}`]);}

export async function registerNotificationRoutes(app:FastifyInstance,config:Config,db:Db,limits:RateLimitStore){
 const listSchema=z.object({cursor:z.string().optional(),limit:z.coerce.number().int().min(1).max(100).default(50),read:z.enum(['true','false']).optional(),category:z.enum(categories).optional(),channel:z.enum(channels).optional(),from:z.iso.datetime().optional(),to:z.iso.datetime().optional()});
 const list=async(r:FastifyRequest,p:FastifyReply)=>{try{const q=listSchema.parse(r.query);if(q.from&&q.to&&q.from>q.to)throw new NotificationError('NOTIFICATION_DATE_RANGE_INVALID','The date range is invalid.',400);const [at,cid]=cursor(q.cursor),x=await db.query(`SELECT n.*,h.read_at FROM notifications n LEFT JOIN notification_read_history h ON h.notification_id=n.id AND h.user_id=$1 WHERE n.recipient_user_id=$1 AND n.authority=$2 AND n.available_at<=now() AND (n.expires_at IS NULL OR n.expires_at>now()) AND ($3::timestamptz IS NULL OR (n.created_at,n.id)<($3,$4)) AND ($5::text IS NULL OR n.category=$5) AND ($6::text IS NULL OR n.channel=$6) AND ($7::timestamptz IS NULL OR n.created_at>=$7) AND ($8::timestamptz IS NULL OR n.created_at<=$8) AND ($9::boolean IS NULL OR (h.read_at IS NOT NULL)=$9) ORDER BY n.created_at DESC,n.id DESC LIMIT $10`,[r.actor!.id,r.actor!.merchantId?'MERCHANT':'PLATFORM',at,cid,q.category??null,q.channel??null,q.from??null,q.to??null,q.read===undefined?null:q.read==='true',q.limit+1]);return {data:x.rows.slice(0,q.limit).map(view),nextCursor:x.rows.length>q.limit?next(x.rows[q.limit-1]):null};}catch(e){return fail(r,p,e)}};
 const unread=async(r:FastifyRequest)=>({count:Number((await db.query(`SELECT count(*) FROM notifications n WHERE n.recipient_user_id=$1 AND n.authority=$2 AND n.channel='IN_APP' AND n.available_at<=now() AND (n.expires_at IS NULL OR n.expires_at>now()) AND NOT EXISTS(SELECT 1 FROM notification_read_history h WHERE h.notification_id=n.id AND h.user_id=$1)`,[r.actor!.id,r.actor!.merchantId?'MERCHANT':'PLATFORM'])).rows[0].count)});
 const mark=async(r:FastifyRequest,p:FastifyReply)=>{try{const nid=id.parse((r.params as any).id),body=z.object({expectedUnread:z.boolean().optional()}).parse(r.body??{}),out=await transaction(db,async c=>idem(c,r,'mark-read',nid,{notificationId:nid,...body},async()=>{const row=(await c.query('SELECT * FROM notifications WHERE id=$1 AND recipient_user_id=$2 AND authority=$3 FOR UPDATE',[nid,r.actor!.id,r.actor!.merchantId?'MERCHANT':'PLATFORM'])).rows[0];if(!row)throw new NotificationError('NOTIFICATION_NOT_FOUND','Notification not found.',404);const inserted=await c.query('INSERT INTO notification_read_history(id,notification_id,user_id) VALUES($1,$2,$3) ON CONFLICT(notification_id,user_id) DO NOTHING RETURNING read_at',[newId('nrh'),nid,r.actor!.id]);if(inserted.rowCount)await evidence(c,r,'NOTIFICATION_READ',nid);return {status:200,body:{id:nid,read:true,readAt:inserted.rows[0]?.read_at??(await c.query('SELECT read_at FROM notification_read_history WHERE notification_id=$1 AND user_id=$2',[nid,r.actor!.id])).rows[0].read_at}};}));return p.code(out.status).send(out.body);}catch(e){return fail(r,p,e)}};
 const readAll=async(r:FastifyRequest,p:FastifyReply)=>{try{const body=z.object({before:z.iso.datetime().optional()}).parse(r.body??{}),out=await transaction(db,async c=>idem(c,r,'mark-all-read','',body,async()=>{const inserted=await c.query(`INSERT INTO notification_read_history(id,notification_id,user_id) SELECT 'nrh_'||replace(gen_random_uuid()::text,'-',''),n.id,$1 FROM notifications n WHERE n.recipient_user_id=$1 AND n.authority=$2 AND n.channel='IN_APP' AND ($3::timestamptz IS NULL OR n.created_at<=$3) ON CONFLICT(notification_id,user_id) DO NOTHING RETURNING notification_id`,[r.actor!.id,r.actor!.merchantId?'MERCHANT':'PLATFORM',body.before??null]);await evidence(c,r,'NOTIFICATIONS_READ_ALL',r.actor!.id,{count:inserted.rowCount});return {status:200,body:{updated:inserted.rowCount}};}));return p.code(out.status).send(out.body);}catch(e){return fail(r,p,e)}};
 const markLimit=rateLimit(limits,config,'notification-mark-read',60,60,r=>`${r.actor?.merchantId??'platform'}:${r.actor?.id}`),allLimit=rateLimit(limits,config,'notification-mark-all',10,60,r=>`${r.actor?.merchantId??'platform'}:${r.actor?.id}`);
 app.get('/v1/notifications',{preHandler:human(db,config,'MERCHANT','notifications.read')},list);app.get('/v1/notifications/unread-count',{preHandler:human(db,config,'MERCHANT','notifications.read')},unread);
 app.post('/v1/notifications/:id/read',{preHandler:[human(db,config,'MERCHANT','notifications.mark_read'),markLimit]},mark);app.post('/v1/notifications/read-all',{preHandler:[human(db,config,'MERCHANT','notifications.mark_read'),allLimit]},readAll);
 app.get('/v1/platform/notifications',{preHandler:human(db,config,'PLATFORM','platform.notifications.read')},list);app.get('/v1/platform/notifications/unread-count',{preHandler:human(db,config,'PLATFORM','platform.notifications.read')},unread);
 app.post('/v1/platform/notifications/:id/read',{preHandler:[human(db,config,'PLATFORM','platform.notifications.read'),markLimit]},mark);app.post('/v1/platform/notifications/read-all',{preHandler:[human(db,config,'PLATFORM','platform.notifications.read'),allLimit]},readAll);
 app.get('/v1/notifications/preferences',{preHandler:human(db,config,'MERCHANT','notifications.manage_preferences')},async r=>{const rows=(await db.query(`SELECT c.category,ch.channel,coalesce(p.enabled,true) enabled,coalesce(p.version,0) version FROM unnest($2::text[]) c(category) CROSS JOIN unnest($3::text[]) ch(channel) LEFT JOIN notification_preferences p ON p.user_id=$1 AND p.category=c.category AND p.channel=ch.channel ORDER BY c.category,ch.channel`,[r.actor!.id,categories,channels])).rows;return {data:rows.map((x:any)=>({category:x.category,channel:x.channel,enabled:x.enabled,effectiveEnabled:mandatory.has(x.category)&&x.channel==='IN_APP'?true:x.enabled,mandatory:mandatory.has(x.category)&&x.channel==='IN_APP',version:Number(x.version),channelAvailable:x.channel==='IN_APP'})),externalDeliveryAvailable:false};});
 app.patch('/v1/notifications/preferences',{preHandler:[human(db,config,'MERCHANT','notifications.manage_preferences'),rateLimit(limits,config,'notification-preferences',10,60,r=>`${r.actor?.merchantId}:${r.actor?.id}`)]},async(r,p)=>{try{const b=z.object({category:z.enum(categories),channel:z.enum(channels),enabled:z.boolean(),expectedVersion:z.number().int().min(0)}).parse(r.body);if(mandatory.has(b.category)&&b.channel==='IN_APP'&&!b.enabled)throw new NotificationError('NOTIFICATION_MANDATORY_PREFERENCE','This mandatory in-app notification cannot be disabled.',409);const out=await transaction(db,async c=>idem(c,r,'preference.update',`${b.category}:${b.channel}`,b,async()=>{const old=(await c.query('SELECT * FROM notification_preferences WHERE user_id=$1 AND category=$2 AND channel=$3 FOR UPDATE',[r.actor!.id,b.category,b.channel])).rows[0];if(Number(old?.version??0)!==b.expectedVersion)throw new NotificationError('NOTIFICATION_PREFERENCE_STALE','The preference version changed.');const row=(await c.query(`INSERT INTO notification_preferences(user_id,merchant_id,category,channel,enabled,version) VALUES($1,$2,$3,$4,$5,1) ON CONFLICT(user_id,category,channel) DO UPDATE SET enabled=excluded.enabled,version=notification_preferences.version+1,updated_at=now() RETURNING *`,[r.actor!.id,r.actor!.merchantId,b.category,b.channel,b.enabled])).rows[0];await evidence(c,r,'NOTIFICATION_PREFERENCES_UPDATED',r.actor!.id,{category:b.category,channel:b.channel,enabled:b.enabled});return {status:200,body:{category:b.category,channel:b.channel,enabled:b.enabled,effectiveEnabled:b.enabled,mandatory:false,version:Number(row.version),channelAvailable:b.channel==='IN_APP',externalDeliveryAvailable:false}};}));return p.code(out.status).send(out.body);}catch(e){return fail(r,p,e)}});
 const deliveryGuard=human(db,config,'PLATFORM','platform.notifications.delivery.read'),searchLimit=rateLimit(limits,config,'notification-delivery-search',60,60,r=>`platform:${r.actor?.id}`);
 app.get('/v1/platform/notification-deliveries',{preHandler:[deliveryGuard,searchLimit]},async()=>({data:(await db.query(`SELECT j.id,j.channel,j.state,j.attempt_count,j.max_attempts,j.available_at,j.lease_expires_at,j.last_error_code,n.merchant_id,n.template_key,n.template_version,r.contact_masked FROM notification_delivery_jobs j JOIN notifications n ON n.id=j.notification_id LEFT JOIN notification_recipients r ON r.notification_id=n.id ORDER BY j.created_at DESC,j.id DESC LIMIT 100`)).rows}));
 app.get('/v1/platform/notification-deliveries/:id',{preHandler:deliveryGuard},async(r,p)=>{const row=(await db.query(`SELECT j.id,j.channel,j.state,j.attempt_count,j.max_attempts,j.available_at,j.lease_expires_at,j.last_error_code,n.merchant_id,n.template_key,n.template_version,r.contact_masked FROM notification_delivery_jobs j JOIN notifications n ON n.id=j.notification_id LEFT JOIN notification_recipients r ON r.notification_id=n.id WHERE j.id=$1`,[(r.params as any).id])).rows[0];return row??p.code(404).send(apiError(r,'NOTIFICATION_DELIVERY_NOT_FOUND','Delivery not found.'));});
 app.get('/v1/platform/notification-templates',{preHandler:human(db,config,'PLATFORM','platform.notifications.templates.read')},async()=>({data:(await db.query('SELECT template_key,version,category,channel,audience_type,required_variables,active,created_at,superseded_at FROM notification_templates ORDER BY template_key,version DESC')).rows}));
 app.post('/v1/platform/notification-deliveries/:id/retry',{preHandler:[human(db,config,'PLATFORM','platform.notifications.delivery.retry'),rateLimit(limits,config,'notification-delivery-retry',10,60,r=>`platform:${r.actor?.id}`),operationalControlGuard(db,'NOTIFICATION_PROCESSING_PAUSED')]},async(r,p)=>{try{const jid=id.parse((r.params as any).id),b=z.object({expectedState:z.enum(['FAILED_RETRYABLE','DEFERRED','DISABLED','SANDBOXED']),reason:z.string().trim().min(3).max(500)}).parse(r.body),out=await transaction(db,async c=>idem(c,r,'delivery.retry',jid,b,async()=>{const j=(await c.query('SELECT * FROM notification_delivery_jobs WHERE id=$1 FOR UPDATE',[jid])).rows[0];if(!j)throw new NotificationError('NOTIFICATION_DELIVERY_NOT_FOUND','Delivery not found.',404);if(j.state!==b.expectedState)throw new NotificationError('NOTIFICATION_DELIVERY_STATE_CONFLICT','The delivery state changed.');if(['EMAIL','SMS'].includes(j.channel))throw new NotificationError('NOTIFICATION_PROVIDER_DISABLED','The external provider is disabled.');await c.query(`UPDATE notification_delivery_jobs SET state='PENDING',available_at=now(),updated_at=now() WHERE id=$1`,[jid]);await evidence(c,r,'NOTIFICATION_DELIVERY_RETRY_REQUESTED',j.notification_id,{deliveryId:jid,channel:j.channel});return {status:200,body:{id:jid,state:'PENDING'}};}));return p.code(out.status).send(out.body);}catch(e){return fail(r,p,e)}});
}

export { categories as notificationCategories, channels as notificationChannels };